Fixed initial page extending on load.
There was a bug with the extending page option, exemplified here: http://gregfranko.com/jquery.tocify.js/.  On initial page load, if you hit the bottom link in the menu, 'Donation', you will see that the page doesn't actually scroll far enough and the extra padding div will not be added until the user has scrolled that far down.  I've corrected this by removing the concept of the removing the concept of the extendPageOffset and added the padding div on load.